catsdogs/requirements.txt
tensorflow==1.12
Pillow
h5py